Replaced non null-safe intent library with android intent library.
